+++ tests/table/mod.rs
@@ -1,0 +1,7 @@
+//! Integration tests for table handling.
+
+mod fixtures;
+mod reflow;
+mod process;
+mod html;
+mod data;
