+++ tests/wrap/lists.rs
@@ -1,0 +1,108 @@
+//! Tests for wrapping of list items.
+//!
+//! Covers plain lists and those containing inline code.
+
+use mdtablefix::process_stream;
+
+#[macro_use]
+#[path = "../prelude/mod.rs"]
+mod prelude;
+use prelude::*;
+
+#[test]
+fn test_wrap_list_item() {
+    let input = lines_vec![
+        r"- This bullet item is exceptionally long and must be wrapped to keep prefix formatting intact.",
+    ];
+    let output = process_stream(&input);
+    assert_wrapped_list_item(&output, "- ", 2);
+}
+
+#[rstest]
+#[case("- ", 3)]
+#[case("1. ", 3)]
+#[case("10. ", 3)]
+#[case("100. ", 3)]
+fn test_wrap_list_items_with_inline_code(#[case] prefix: &str, #[case] expected: usize) {
+    let input = lines_vec![format!(
+        "{prefix}`script`: A multi-line script declared with the YAML `|` block style. The entire \
+         block is passed to an interpreter. If the first line begins with `#!`, Netsuke executes \
+         the script verbatim, respecting the shebang."
+    )];
+    let output = process_stream(&input);
+    assert_wrapped_list_item(&output, prefix, expected);
+}
+
+#[test]
+fn test_wrap_preserves_inline_code_spans() {
+    let input = lines_vec![
+        "- `script`: A multi-line script declared with the YAML `|` block style. The entire block \
+         is passed to an interpreter. If the first line begins with `#!`, Netsuke executes the \
+         script verbatim, respecting the shebang.",
+    ];
+    let output = process_stream(&input);
+    assert_wrapped_list_item(&output, "- ", 3);
+}
+
+#[test]
+fn test_wrap_multi_backtick_code() {
+    let input = lines_vec![
+        "- ``cmd`` executes ```echo``` output with ``json`` format and prints results to the \
+         console",
+    ];
+    let output = process_stream(&input);
+    assert_wrapped_list_item(&output, "- ", 2);
+}
+
+#[test]
+fn test_wrap_multiple_inline_code_spans() {
+    let input = lines_vec![
+        "- Use `foo` and `bar` inside ``baz`` for testing with additional commentary to exceed \
+         wrapping width",
+    ];
+    let output = process_stream(&input);
+    assert_wrapped_list_item(&output, "- ", 2);
+}
+
+#[test]
+fn test_wrap_long_inline_code_item() {
+    let input = lines_vec![concat!(
+        "- `async def on_unhandled(self, ws: WebSocketLike, message: Union[str, bytes])`:",
+        " A fallback handler for messages that are not dispatched by the more specific",
+        " message handlers. This can be used for raw text/binary data or messages that",
+        " don't conform to the expected structured format."
+    )];
+    let output = process_stream(&input);
+    assert_wrapped_list_item(&output, "- ", 4);
+    assert!(
+        output
+            .first()
+            .expect("wrapped output should contain at least one line")
+            .ends_with("`:")
+    );
+}
+
+#[test]
+fn test_wrap_future_attribute_punctuation() {
+    let input = lines_vec![concat!(
+        "- Test function (`#[awt]`) or a specific `#[future]` argument ",
+        "(`#[future(awt)]`), tells `rstest` to automatically insert `.await` ",
+        "calls for those futures."
+    )];
+    let output = process_stream(&input);
+    assert_eq!(
+        output,
+        vec![
+            "- Test function (`#[awt]`) or a specific `#[future]` argument".to_string(),
+            "  (`#[future(awt)]`), tells `rstest` to automatically insert `.await` calls for".to_string(),
+            "  those futures.".to_string(),
+        ]
+    );
+}
+
+#[test]
+fn test_wrap_short_list_item() {
+    let input = lines_vec!["- short item"];
+    let output = process_stream(&input);
+    assert_eq!(output, input);
+}
