+++ tests/wrap/mod.rs
@@ -1,0 +1,8 @@
+//! Test suite for wrapping functionality.
+
+mod paragraphs;
+mod lists;
+mod footnotes;
+mod blockquotes;
+mod hard_breaks;
+mod cli;
