+++ tests/wrap/paragraphs.rs
@@ -1,0 +1,77 @@
+//! Tests for paragraph wrapping behaviour.
+//!
+//! Verifies wrapping of plain text paragraphs and handling of links.
+
+use mdtablefix::process_stream;
+
+#[macro_use]
+#[path = "../prelude/mod.rs"]
+mod prelude;
+use prelude::*;
+
+#[test]
+fn test_wrap_paragraph() {
+    let input = lines_vec![
+        "This is a very long paragraph that should be wrapped at eighty columns so it needs to \
+         contain enough words to exceed that limit.",
+    ];
+    let output = process_stream(&input);
+    assert!(output.len() > 1);
+    assert!(output.iter().all(|l| l.len() <= 80));
+}
+
+#[test]
+fn test_wrap_paragraph_with_long_word() {
+    let long_word = "a".repeat(100);
+    let input = lines_vec![&long_word];
+    let output = process_stream(&input);
+    assert_eq!(output.len(), 1);
+    assert_eq!(output[0], long_word);
+}
+
+#[test]
+fn test_wrap_paragraph_with_link() {
+    let input = lines_vec![concat!(
+        "**Wireframe** is an experimental Rust library that simplifies building",
+        " servers and clients for custom binary protocols. The design borrows ",
+        "heavily from [Actix Web](https://actix.rs/) to provide a familiar, ",
+        "declarative API for routing, extractors, and middleware.",
+    )];
+    let output = process_stream(&input);
+    assert!(
+        output
+            .iter()
+            .any(|line| line.contains("[Actix Web](https://actix.rs/)")),
+        "link should not be broken across lines",
+    );
+}
+
+#[test]
+fn test_wrap_paragraph_with_image_link() {
+    let input = lines_vec![concat!(
+        "Here is an image ![logo](https://example.com/logo.png) embedded in ",
+        "a sentence that should wrap without splitting the link.",
+    )];
+    let output = process_stream(&input);
+    assert!(
+        output
+            .iter()
+            .any(|line| line.contains("![logo](https://example.com/logo.png)")),
+        "image link should not be broken across lines",
+    );
+}
+
+#[test]
+fn test_wrap_paragraph_with_nested_link() {
+    let input = lines_vec![concat!(
+        "Check [docs](https://example.com/rust(nightly)/guide) for details on",
+        " nightly features and usage.",
+    )];
+    let output = process_stream(&input);
+    assert!(
+        output
+            .iter()
+            .any(|line| line.contains("(https://example.com/rust(nightly)/guide)")),
+        "link with nested parentheses should remain intact",
+    );
+}
